move components to ui folder
