+++ nym-browser-extension/src/components/ui/Logo/index.tsx
