+++ nym-browser-extension/src/components/ui/LogoWithText/index.tsx
