+++ nym-browser-extension/src/components/ui/Title/index.tsx
@@ -1,5 +1,5 @@
-import { Typography } from '@mui/material';
 import React from 'react';
+import { Typography } from '@mui/material';
 
 const FONT_SIZE = 22;
 const FONT_WEIGHT = 400;
